+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Uses save state file version 13.
     - `smpc-asia.bin`: Korea, Taiwan -- SMPC area codes 2, 6
     - `smpc-other.bin`: Other (invalid) SMPC area codes
     - The old `smpc.bin` will be automatically migrated to these files as you use IPL ROMs for each region.
+- Debugger: Added RBG0 and RBG1 line color single stack views to the VDP2 debug overlay.
 - Input: Added option to constrain mouse cursor to window in system cursor mode.
 - Input: Convert 3D Control Pad analog stick to D-Pad inputs when in digital mode.
 - Input: Graduate Virtua Gun to stable feature.